Raise exception if no data can be plotted
I'm currently trying to get a run done in my local setup where all validations fail due to a too low amount of data. Locally I adapted QA4SM to also generate graphs in this case. This leads to only NaNs in the result. This leads to an error here:

https://github.com/awst-austria/qa4sm-reader/blob/583a084392870b81d730523ae27a2bf266e115f0/src/qa4sm_reader/plotter.py#L906-L912

Concatenating fails, because _yield_values does not return anything, and `values` stays an empty list.

https://github.com/awst-austria/qa4sm-reader/blob/583a084392870b81d730523ae27a2bf266e115f0/src/qa4sm_reader/plotter.py#L366

As a result, the only plot that is generated is the one for #observations. Raising the proper error here will make it possible for the plotter to handle this error, and consequentially we would also get the error code plots.